fixed fgThread loop. rudder and throttle now working!

+++ src/main/java/com/example/remotejoystick17/view_model/ViewModel.java
@@ -31,7 +31,7 @@ public class ViewModel extends Observable implements Observer {
         model.rudder(r);
     }
 
-    public void setThrottle(float t) {
+    public void setThrottle(double t) {
         model.throttle(t);
     }
